Fix nif salt returning 64-byte binary with unexpected data
nif was allocating an Erlang binary of 64 bytes, writing the salt to it
(shorter than 64 bytes) and then returning the salt plus whatever data
was set to that memory previously.

Since it was allocated properly it shouldn't result in segfaults but
it's simply not data we want there.

problem identified and solved by @mtornwall